Refactor EmbeddingsModel
